+++ api/person.py
@@ -1,10 +1,12 @@
 from flask_restful import Resource
-from flask import jsonify,request
+from flask import jsonify,request,session
 from shared_db import db
 
 from models.models import Person
 
 class PersonResource(Resource):
+
+    # TODO k tomuto treba debatu (profil alebo jak?/ admin?)
     def get(self,id = None):
         if id is None:
             person = Person.query.all()
@@ -26,8 +28,12 @@ class PersonResource(Resource):
 
             return jsonify(person)
 
-
+    # Register
+    # Can be done if no session is active
     def post(self,id= None):
+        if session["user_id"]:
+            return "nenene"
+
         #osetrit ze mail uz je zadany + aj username
         email     = request.form.get("email")
         user_type = request.form.get("user_type")
@@ -47,13 +53,20 @@ class PersonResource(Resource):
         db.session.add(person)
         db.session.commit()
 
-
+    # Remove user
+    # Can be done by Admin
     def delete(self,id):
+        if not session['user_type'] == 5:
+            return "nenenene"
         Person.query.filter_by(id=id).delete()
         db.session.commit()
 
-
+    # Edit user
+    # Can be done by Admin and User(can edit his info) TODO moze user editovat vsetko? (user_type urcite nope)
     def put(self,id):
+        if not (session['user_id'] == id or session['user_type'] == 5):  # is the right person logged // admin
+            return "nenenene"
+
         person = Person.query.filter_by(id=id).first()
 
         if not person:
